Move CoreDataFeedStore extension to production scope
